requirements.txt
Cython==0.29
ptest==3.9.3
PyHamcrest==1.9.0
pyjnius==1.1.3
